Feature: implementing new identifier for participant data

+++ src/service/participant_data/get.class.php
@@ -27,10 +27,30 @@ class get extends \cenozo\service\downloadable
    */
   protected function get_downloadable_public_name()
   {
+    $identifier_class_name = lib::get_class_name( 'database\identifier' );
+    $participant_identifier_class_name = lib::get_class_name( 'database\participant_identifier' );
+    $setting_manager = lib::create( 'business\setting_manager' );
+
+    $identifier = NULL;
+    $participant_data_identifier = $setting_manager->get_setting( 'general', 'participant_data_identifier' );
+    if( !is_null( $participant_data_identifier ) )
+    {
+      $db_identifier = $identifier_class_name::get_unique_record( 'name', $participant_data_identifier );
+      if( !is_null( $db_identifier ) )
+      {
+        $db_participant_identifier = $participant_identifier_class_name::get_unique_record(
+          array( 'identifier_id', 'participant_id' ),
+          array( $db_identifier->id, $this->db_participant->id )
+        );
+        if( !is_null( $db_participant_identifier ) ) $identifier = $db_participant_identifier->value;
+      }
+    }
+
     $db_participant_data = $this->get_leaf_record();
     $db_study_phase = $db_participant_data->get_study_phase();
     return sprintf(
-      '%s %s %s %s.%s',
+      '%s%s %s %s %s.%s',
+      is_null( $identifier ) ? '' : sprintf( '%s ', $identifier ),
       $db_study_phase->get_study()->name,
       strtoupper( $db_study_phase->code ),
       $db_participant_data->category,
@@ -73,8 +93,12 @@ class get extends \cenozo\service\downloadable
     // clean up by deleting temporary files
     if( $this->get_argument( 'download', false ) && !is_null( $this->db_participant ) )
     {
-      $filename = $this->get_leaf_record()->get_filename( $this->db_participant );
-      if( file_exists( $filename ) ) unlink( $filename );
+      $db_participant_data = $this->get_leaf_record();
+      if( is_null( $db_participant_data ) )
+      {
+        $filename = $db_participant_data->get_filename( $this->db_participant );
+        if( file_exists( $filename ) ) unlink( $filename );
+      }
     }
   }
